+++ allshoppings-rest-api/src/main/java/mobi/allshoppings/lock/LockHelper.java
@@ -6,7 +6,8 @@ import mobi.allshoppings.exception.ASException;
 
 public interface LockHelper {
 
-	void deviceMessageLock(String deviceId, Integer scope, String campaignActivityId, Date fromDate, long duration, String subEntityId, Integer subEntityKind) throws ASException;
-	void clearLocks(String deviceId, Integer scope, String campaignActivityId) throws ASException;
+	void deviceMessageLock(String deviceId, byte scope, String campaignActivityId, Date fromDate,
+			long duration, String subEntityId, int subEntityKind) throws ASException;
+	void clearLocks(String deviceId, byte scope, String campaignActivityId) throws ASException;
 	
 }
